+++ runtime/src/commonMain/kotlin/pbandk/json/MessageExtensions.kt
@@ -1,18 +1,18 @@
 package pbandk.json
 
 import pbandk.ExperimentalProtoJson
+import pbandk.Export
 import pbandk.InvalidProtocolBufferException
 import pbandk.Message
 import pbandk.decodeWith
 import pbandk.internal.json.JsonMessageEncoder
 import pbandk.internal.json.JsonMessageDecoder
-import kotlin.js.JsExport
 
 /**
  * Encode this message to a String using the protocol buffer JSON encoding.
  */
 @ExperimentalProtoJson
-@JsExport
+@Export
 public fun <T : Message> T.encodeToJsonString(jsonConfig: JsonConfig = JsonConfig.DEFAULT): String =
     JsonMessageEncoder(jsonConfig).also { it.writeMessage(this) }.toJsonString()
 
@@ -20,7 +20,7 @@ public fun <T : Message> T.encodeToJsonString(jsonConfig: JsonConfig = JsonConfi
  * Decode a JSON protocol buffer message from [data].
  */
 @ExperimentalProtoJson
-@JsExport
+@Export
 @Throws(InvalidProtocolBufferException::class)
 public fun <T : Message> Message.Companion<T>.decodeFromJsonString(
     data: String,