+++ runtime/src/wasmJsMain/kotlin/pbandk/Export.wasmJs.kt
@@ -1,0 +1,8 @@
+@file:Suppress("EXPECT_ACTUAL_CLASSIFIERS_ARE_IN_BETA_WARNING")
+
+package pbandk
+
+@Target(AnnotationTarget.CLASS, AnnotationTarget.PROPERTY, AnnotationTarget.FUNCTION, AnnotationTarget.FILE)
+public actual annotation class Export
+
+public actual typealias JsName = kotlin.js.JsName